requirements.txt
pynput
google-genai
pyinstaller